refactor: remove re-exports from chips index file
Eliminated re-exports from the chips index file to improve module structure. Updated all imports to use specific component paths, enhancing code clarity and reducing potential ambiguity.
