--- apps/nextjs/src/components/features/chips/index.ts
@@ -1,5 +1,0 @@
-export * from './assignee-chip';
-export * from './invited-teammate-chip';
-export * from './priority-chip';
-export * from './project-chip';
-export * from './tag-chip';
