+++ apps/nextjs/src/components/features/modals/share-project-modal/body/share/send-form.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useState } from 'react';
-import { InvitedTeammateChip } from '@/components/features/chips';
+import { InvitedTeammateChip } from '@/components/features/chips/invited-teammate-chip';
 import { InviteProjectTeammateMenu } from '@/components/features/menus';
 import { Flex } from '@/components/ui/flex';
 import { Input as AtomsInput } from '@/components/ui/input';
