+++ apps/nextjs/src/components/features/tasks/tasks-board/tasks-board-list-item/priority/priority.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { PriorityChip } from '@/components/features/chips';
+import { PriorityChip } from '@/components/features/chips/priority-chip';
 import type { FlexProps } from '@/components/ui/flex';
 import { useTask } from '@/store/entities/task';
