+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-projects/list-item.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { ProjectChip } from '@/components/features/chips';
+import { ProjectChip } from '@/components/features/chips/project-chip';
 import { useProjectTask } from '@/store/entities/project-task';
 
 type Props = {
