+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-tags/input/input.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useMemo, useRef, useState } from 'react';
-import { TagChip } from '@/components/features/chips';
+import { TagChip } from '@/components/features/chips/tag-chip';
 import { TagMenu } from '@/components/features/menus';
 import { Flex } from '@/components/ui/flex';
 import { Input as AtomsInput } from '@/components/ui/input';
